% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mul_lin_reg.R
\name{lin_reg}
\alias{lin_reg}
\title{Create a linear regression model for the given input}
\usage{
lin_reg(formula, data)
}
\arguments{
\item{formula}{a formula in the form y ~ x where y is the response variable and x is the sum of all of the covariates}

\item{data}{a data set containing x and y}
}
\value{
Coefficient estimates and their standard errors and p-values for the multiple linear regression model
}
\description{
Create a linear regression model for the given input
}
\examples{
library(EGeis625)
lin_reg(mpg~disp + cyl + hp + drat + wt + as.factor(gear), mtcars)
}
